sgp4: fix signed integer overflow in epochStar to Julian date conversion
epochStar is a signed integer representing the time offset from Monday
00:00 UTC. Negative values are valid and indicate an epoch in the
previous week. Multiplying by 1000UL silently converted the signed
int to unsigned long, wrapping negative values to large positive ones.

Cast to int64_t before the multiplication to preserve sign and avoid
overflow on both 32-bit and 64-bit (test) targets.

+++ app/src/modules/sgp4/sgp4_pass_predict.c
@@ -266,7 +266,7 @@
 	 LOG_DBG("weekday: %d unix_time_ms1: %lld", weekday, unix_time_ms);
 
 	 /* Deduct 86400 * weekday to get the unix time of last monday */
-	 unix_time_ms -= (weekday * 86400000UL);
+	 unix_time_ms -= ((int64_t)weekday * 86400000LL);
 	 LOG_DBG("unix_time_ms2: %lld", unix_time_ms);
 	 datetime_from_unix_time_ms(unix_time_ms, &dt_mon_midnight);
 
@@ -277,7 +277,7 @@
 
 	 unix_time_ms = datetime_to_ts(&dt_mon_midnight);
 	 LOG_DBG("unix_time_ms3: %lld", unix_time_ms);
-	 unix_time_ms += (epochStar * 1000UL);
+	 unix_time_ms += ((int64_t)epochStar * 1000LL);
 	 datetime_from_unix_time_ms(unix_time_ms, &dt_mon_midnight);
 	 jd_from_unix_time_ms(unix_time_ms, jdsatepoch, jdsatepochF);
